Move message post-processing closer to place of definition
